add Vladimir Bordiyan sum double

+++ src/Calculator.java
@@ -17,6 +17,6 @@ public class Calculator {
     public static int addInt (int a,int b){
         return a+b;
     }
-    public static double sum (int a,int b){
+    public static double sum (double a,double b){
         return a+b;
 }
